@@ -1,17 +1,17 @@
 name: simple_bloc
 description: A simple alternative to the BLoC implementation created in Dimension C-137
-version: 1.1.9
+version: 1.2.0
 homepage: https://github.com/everton-e26/simple_bloc
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
-  rxdart: ^0.24.1
-  provider: ^4.1.3
+  rxdart: ^0.27.2
+  provider: ^6.0.1
 
 dev_dependencies:
   flutter_test:
