@@ -6,8 +6,8 @@ import 'package:simple_bloc/simple_bloc.dart';
 
 /// Base bloc class
 abstract class Bloc {
-  BuildContext _context;
-  BuildContext get context => _context;
+  BuildContext? _context;
+  BuildContext? get context => _context;
   final _onDisposeController = BehaviorSubject<Bloc>();
   void initialize() {}
 
@@ -27,20 +27,20 @@ abstract class Bloc {
 }
 
 class BlocInstanceManager {
-  final List<BlocBuilder> blocs;
-  final Map<String, Bloc> _blocs = {};
+  final List<BlocBuilder>? blocs;
+  final Map<String, Bloc?> _blocs = {};
 
   BlocInstanceManager(this.blocs);
 
-  T of<T extends Bloc>(BuildContext context) {
+  T? of<T extends Bloc?>(BuildContext context) {
     String type = T.toString();
-    T bloc;
+    T? bloc;
     if (_blocs.containsKey(type)) {
-      bloc = _blocs[type];
+      bloc = _blocs[type] as T?;
     } else {
-      BlocBuilder b = blocs.firstWhere((b) => b.build is T Function());
-      bloc = b.build();
-      bloc._context = context;
+      BlocBuilder b = blocs!.firstWhere((b) => b.build is T Function());
+      bloc = b.build() as T?;
+      bloc!._context = context;
       bloc.initialize();
       _blocs.putIfAbsent(type, () => bloc);
     }
@@ -49,7 +49,7 @@ class BlocInstanceManager {
 
   void dispose() {
     _blocs.forEach((key, bloc) {
-      bloc._dispose();
+      bloc!._dispose();
     });
   }
 }
